% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scapp.R
\name{scapp}
\alias{scapp}
\title{app demo for SingleR application}
\usage{
scapp()
}
\description{
app demo for SingleR application
}
\note{
We work just with the MuraroPancreasData, leaving
more general upload capability as a project.
}
